graded group set
