+++ packages/perseus/src/prompt-types.ts
@@ -4,7 +4,7 @@ import type {DropdownPromptJSON} from "./widget-ai-utils/dropdown/dropdown-ai-ut
 import type {ExplanationPromptJSON} from "./widget-ai-utils/explanation/explanation-ai-utils";
 import type {ExpressionPromptJSON} from "./widget-ai-utils/expression/expression-ai-utils";
 import type {GradedGroupPromptJSON} from "./widget-ai-utils/graded-group/graded-group-ai-utils";
-import type {GradedGroupSetPromptJSON} from "./widget-ai-utils/graded-group-set/prompt-utils";
+import type {GradedGroupSetPromptJSON} from "./widget-ai-utils/graded-group-set/graded-group-set-ai-utils";
 import type {GrapherPromptJSON} from "./widget-ai-utils/grapher/prompt-utils";
 import type {GroupPromptJSON} from "./widget-ai-utils/group/prompt-utils";
 import type {ImagePromptJSON} from "./widget-ai-utils/image/prompt-utils";
